move samples/interactive-tutorials/product/setup to move samples/interactive-tutorials/product/setup_product
