+++ samples/interactive-tutorials/product/setup_product/products_create_bigquery_table.py
