+++ samples/interactive-tutorials/product/setup_product/products_create_gcs_bucket.py
