+++ samples/interactive-tutorials/product/setup_product/setup_cleanup.py
